moving kubernetes config templates to a package
